+++ R/RowLinkedMatrix.R
@@ -244,6 +244,23 @@ index.RowLinkedMatrix <- function(x, i = NULL, ...) {
 }
 
 
+#' Converts an Object to a RowLinkedMatrix Object.
+#'
+#' @param x An object to convert to a [RowLinkedMatrix-class] object.
+#' @param ... Additional arguments.
+#' @return A [RowLinkedMatrix-class] object.
+#' @export
+as.RowLinkedMatrix <- function(x, ...) {
+    UseMethod("as.RowLinkedMatrix")
+}
+
+
+#' @export
+as.RowLinkedMatrix.list <- function(x, ...) {
+    do.call("RowLinkedMatrix", x, ...)
+}
+
+
 #' @rdname ColumnLinkedMatrix-class
 #' @export RowLinkedMatrix
 #' @exportClass RowLinkedMatrix