+++ NEWS.md
@@ -1,5 +1,8 @@
 # LinkedMatrix 1.3.0.9000
 
+* Add generic `as.ColumnLinkedMatrix` and `as.RowLinkedMatrix` for easy
+  creation of `LinkedMatrix` objects from `list`s of matrix-like objects
+  without `do.call`.
 * Warn if rownames (for `ColumnLinkedMatrix`) or colnames (for
   `RowLinkedMatrix`) do not match.
 * Add examples.